Revert "[no-release-notes] Revert "[no-release-notes] Reorganize query script tests""
